requirements.txt
Django
djangorestframework
psycopg2-binary
bcrypt
PyJWT
dotenv
djangorestframework-simplejwt
drf-spectacular